requirements.txt
google-cloud-bigquery==3.3.2
pandas~=1.4.3
db-dtypes==1.0.3
PyYAML~=6.0
fastapi==0.98.0
uvicorn==0.22.0
expiringdict==1.2.2
google-cloud-logging==3.2.5
google-cloud-error-reporting==1.9.1
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-exporter-gcp-trace==1.4.0
pytest
